+++ tapo-py/src/api_client.rs
@@ -2,7 +2,7 @@ use pyo3::prelude::*;
 use tapo::ApiClient;
 
 use crate::errors::ErrorWrapper;
-use crate::handlers::{PyPlugEnergyMonitoringHandler, PyPlugHandler};
+use crate::handlers::{PyGenericDeviceHandler, PyPlugEnergyMonitoringHandler, PyPlugHandler};
 
 #[pyclass(name = "ApiClient")]
 pub struct PyApiClient {
@@ -15,6 +15,17 @@ impl PyApiClient {
     pub fn new(tapo_username: String, tapo_password: String) -> Result<Self, ErrorWrapper> {
         let client = ApiClient::new(tapo_username, tapo_password)?;
         Ok(Self { client })
+    }
+
+    pub fn generic_device<'a>(&'a self, ip_address: String, py: Python<'a>) -> PyResult<&'a PyAny> {
+        let client = self.client.clone();
+        pyo3_asyncio::tokio::future_into_py(py, async move {
+            let handler = client
+                .generic_device(ip_address)
+                .await
+                .map_err(ErrorWrapper)?;
+            Ok(PyGenericDeviceHandler::new(handler))
+        })
     }
 
     pub fn p100<'a>(&'a self, ip_address: String, py: Python<'a>) -> PyResult<&'a PyAny> {