+++ tapo-py/src/handlers.rs
@@ -1,5 +1,7 @@
+mod generic_device_handler;
 mod plug_energy_monitoring_handler;
 mod plug_handler;
 
+pub use generic_device_handler::*;
 pub use plug_energy_monitoring_handler::*;
 pub use plug_handler::*;