+++ tapo-py/src/handlers/generic_device_handler.rs
@@ -1,0 +1,66 @@
+use std::sync::Arc;
+
+use pyo3::prelude::*;
+use tapo::GenericDeviceHandler;
+use tokio::sync::Mutex;
+
+use crate::errors::ErrorWrapper;
+
+#[derive(Clone)]
+#[pyclass(name = "GenericDeviceHandler")]
+pub struct PyGenericDeviceHandler {
+    handler: Arc<Mutex<GenericDeviceHandler>>,
+}
+
+impl PyGenericDeviceHandler {
+    pub fn new(handler: GenericDeviceHandler) -> Self {
+        Self {
+            handler: Arc::new(Mutex::new(handler)),
+        }
+    }
+}
+
+#[pymethods]
+impl PyGenericDeviceHandler {
+    pub fn refresh_session<'a>(&'a self, py: Python<'a>) -> PyResult<&'a PyAny> {
+        let handler = self.handler.clone();
+        pyo3_asyncio::tokio::future_into_py(py, async move {
+            handler
+                .lock()
+                .await
+                .refresh_session()
+                .await
+                .map_err(ErrorWrapper)?;
+            Ok(())
+        })
+    }
+
+    pub fn on<'a>(&'a self, py: Python<'a>) -> PyResult<&'a PyAny> {
+        let handler = self.handler.clone();
+        pyo3_asyncio::tokio::future_into_py(py, async move {
+            handler.lock().await.on().await.map_err(ErrorWrapper)?;
+            Ok(())
+        })
+    }
+
+    pub fn off<'a>(&'a self, py: Python<'a>) -> PyResult<&'a PyAny> {
+        let handler = self.handler.clone();
+        pyo3_asyncio::tokio::future_into_py(py, async move {
+            handler.lock().await.off().await.map_err(ErrorWrapper)?;
+            Ok(())
+        })
+    }
+
+    pub fn get_device_info<'a>(&'a self, py: Python<'a>) -> PyResult<&'a PyAny> {
+        let handler = self.handler.clone();
+        pyo3_asyncio::tokio::future_into_py(py, async move {
+            let device_info = handler
+                .lock()
+                .await
+                .get_device_info()
+                .await
+                .map_err(ErrorWrapper)?;
+            Ok(device_info)
+        })
+    }
+}